Use proper integration testing for MembersService.
