+++ member/src/main/scala/com/improving/app/member/api/MemberServiceImpl.scala
@@ -15,7 +15,7 @@ import scala.language.postfixOps
 class MemberServiceImpl(implicit val system: ActorSystem[_]) extends MemberService {
 
   implicit val ec: ExecutionContext = system.executionContext
-  implicit val timeout: Timeout = Timeout(5 seconds)
+  implicit val timeout: Timeout = Timeout(20 seconds)
 
   //Create a new member
   val sharding: ClusterSharding = ClusterSharding(system)
